requirements.txt
tensorflow>=2.5
numpy
matplotlib
Pillow
transformers
torch